update implements according to suggestion
